Move JIT backend to a different file
